wait this is actually so cool

+++ OpenGL/src/physics/PhysicsEngine.h
@@ -9,7 +9,7 @@ namespace Atlas {
 	class PhysicsEngine {
 	public:
 		static void Initialize();
-		static void Update(float deltaT);
+		static void Update(double deltaT);
 		static btCollisionObject* AddPhysicsBody(btCollisionShape* shape, btTransform& transformation, float mass);
 		static void Recalculate(btCollisionObject* object);
 		static void DrawDebug();
